requirements.txt
flask>=3.0.0
yfinance>=0.2.36
pytz>=2024.1
alpaca-py>=0.26.0
